refactor(c1-b3): split env filter into focused modules
